[IDLE-313] 요양보호사 메인화면 메인 코디네이터 구현
